Hide "Help" popover on Escape
Part of story #18337 markdown replace text

How to test:
$ (cd plugins/tracker/script/lib/rich-text-editor && npm run build)
$ (cd plugins/tracker && npm run build)
$ (cd plugins/testmanagement && npm run build)

Run the test in the following places:
- When editing an artifact, in text fields.
- When editing a test definition, in steps field.

Open the "Syntax help" button. Hit "Esc" on your keyboard. The "Syntax
help" popover should hide.

Note: The Artifact modal syntax help is not handled. It will be a
dedicated change, as there is no hide() method in TLP popovers.

Change-Id: I6922c9194c7bf0718c5047809843303bee5b5553