Display dependencies in Roadmap
This displays all links, whatever their nature. (Later, user will be able
to select which type of link must be displayed.)

Color of the arrow is fixed to "dimmed", it will be adjusted later.

As of today, the dimensions of a task (position from the left + width)
might be computed more than once:
* once for the bar itself
* once for each dependency the task is part of
It will be improved in a dedicated changeset.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I6f74d3695f92684558bbee0b87a25d9810eff6f5