request #20149: Use ESBuild to transpile and minimize the JS/TS and CSS code
This contribution replaces terser-webpack-plugin/Terser, css-minimizer-webpack-plugin/cssnano
and babel-loader/Babel by esbuild-loader/ESBuild.

This is a trade-off between build performances and asset sizes. In our
situation where the first load does not matter that much, it is probably
acceptable (and it's not impossible we gain back those bytes in the
future).

The persistent file caching of Webpack has been disabled. With this new
setup and the concurrent build of the packages it does not seem to bring
a lot of improvements but it increases the I/O load. We probably can
gain a bit from it but it requires more tweaking that might not be
worth it right now.

There is a small code modification in the document plugin to workaround
an edge case (fixed upstream) regarding the minification of the object
spread.

Change-Id: I3c9fe7af77033f9a40cab1b0dbdb00356f491761