Do not change tracker id to xsd:ID
Fixing a bug on the xml definition can have impacts at the other side of
the ball of yarn.

Switch back the type to string. This means that we cannot use standard
widget xml definition to reference trackers.

The new syntax of the roadmap widget is:

```
  <widget name="plugin_roadmap_project_widget">
    <preference name="roadmap">
      <value name="title">My roadmap</value>
      <value name="tracker_id">T754</value>
    </preference>
  </widget>
```

No functional changes, the rest tests still pass.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I968569483ab2ce575ff3da8f96ce06b7ade0ca38