User can choose quarter timescale
When tasks span on long period of time, user may want to change the
timescale so that information is more readable.

This allows us to put time units tooltip on top instead of bottom of the
header.

Known bug: at specific widths of the viewport, an horizontal scrollbar
appears for timescale:week, whereas it is not the case for
timescale:month. I suggest that we investigate this later.

getBeginningOfPeriod (and getEndOfPeriod) will be deduplicated later.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I5f497b0b15189d470bbaff554e51fa98f4ca4189