Be able to close a campaign in REST
This is part of request #20088 Improve the management of closed campaigns

You are now able to close a campaign using the REST route
PATCH testmanagement_campaigns/id with the following body:

{
  "change_status": "closed"
}

This uses the same behaviour defined for Kanban. We find the first not
open value defined in the campaign tracker and we apply it to the
campaign.

Automated REST tests will be added in the next patch with the ability to
open a campaign.

Change-Id: I19b0b4259557204fec2d3a0b67ad4d80aedf5ffd