Close Artifact modal "Help" popover on Escape
Part of story #18337 markdown replace text

How to test:
$ (cd src/ && npm run build)
$ (cd plugins/agiledashboard && npm run build)

Edit or create a new artifact with a text field. Switch to "Markdown"
format. Click on the "Help" button. Press "Escape" key on your keyboard
=> The "Help" popover should close.

When you switch format to HTML or close the modal, the event listener
should be removed from document.

Change-Id: I8888d5c8583ecc333181b3341bcf3524e0974355