Route POST commit status should be public
The route has been defined as protected in recent development.
It causes external requests to fail as they are not authenticated
when they pass the CI token instead of an access key.

part of: story #18320 native support of pull requests in tuleap branch source jenkins plugin

How to test:
- In the api explorer, query the route using the ci token
--> it works

- Make sure you belong to a user group authorized to set build statuses
- Remove the ci token from the payload
--> it works

- Remove yourself from the authorized group
--> 403

- Using Curl, query the route using the ci token, MAKE SURE YOU DO NOT
  PASS YOUR ACCESS KEY IN THE HEADERS
--> It works

- Remove the ci token from the payload and add your access key in the
  headers
--> 403 (since you removed yourself from the authorized groups in a
previous step)

- Add yourself to an authorized group
--> It works

Change-Id: I69b00dc24b24d93cc8b6bfb6930c1760ad300fa6