User can choose which type of link to display
Given a Task A has a link to Task B with the nature "depends_on"
When the user first display the widget
Then no dependency is displayed
And the user can choose to display the links "Depends on" (or other type
of links).

Note: Links without nature are labelled "", like it is done in the
tracker interface. However I think that we should probably put a
distinctive label. IMHO this can be released as is and enhanced later in
a dedicated commit after discussion with Design Team.

Note: Since the beginning of the story, we only display "forward"
dependency. Once released and put into real users hand, this patch will
highlight if we need also "reverse" links.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: Ic39b2fc55910d4e96eaf3a3d65412b443d8a98f7