Prevent StepDef field to be updated via REST calls
Issue is present since the beginning but highlighted by story #19304:
now we return the definition of the steps in the artifact
representation. However, since the artifact modal send back the artifact
representation in the PUT, it tries to update the field.

Fix request #20156: Updating description in test modal clears the steps

Change-Id: I7c16f280487386f5c531d83bac37f4cfa54c99e8