Move Babel configuration for Jest outside the Webpack configuration
Now that the Webpack configuration does not use Babel, it does not make
a lot of sense to keep the Babel configuration for Jest in there.

This contribution also modernizes the Babel configuration for our
current environment.

Part of request #20149: Use ESBuild to transpile and minimize the JS/TS and CSS code

Change-Id: I47fb53dc0636dc1498814b9eacd26e47a0434c5d