request #14917 Remove admin delegation widgets
They are broken since ages, big warnings were added to inform about
futur removal. Nobody complained, kill it with fire.

Despite the removal of the plugin, the data are not purged from the
DB as we tend to keep data as much as possible. The admins will be
warned in the deployed guide about the purge if they want to run it
by hand.

At uninstall of the plugin, the widgets were purged. However in this
context (forge upgrade) we cannot rely on production code and
duplicating the SQL queries insider the forge upgrade is error prone and
probably not worth it. Keep in mind that nobody bothered about the
removal so it probably means that nobody are using them anyway.

Change-Id: I7b723740755820e75489012a14dbba83d4078371