request #20896 : Improving header accessibility
User should be able to navigate the header menu (+new, user nav…) with the keyboard.
Popovers should not take focus and be tabbable when not open.
They should also close when hitting the escape key, similarly to dropdowns.
Trigger buttons should be real buttons so that they can be "clicked" with the keyboard.

In FlamingParrot, `new` and `user nav` popovers are now made using Tuleap popovers
rather than Bootstrap dropdowns.

How to test:
* When navigating in the header, tab does not go through popovers
  content if they are not open (`+New` and `user` menus).
* You can open the popovers by hitting the `enter` key on the
  trigger buttons.
* When a popover is open, you can tab through it.
* When hitting the `escape` key, the popover closes.

Change-Id: I4609b47986d2a6e6a7c6a58d0b24afe2ea053392