Improve PATCH REST route project/$id/program_backlog to move feature from PI to backlog
This is part of story #20111 change features priority

This patch has goal to improve REST route patch `project/$id/program_backlog`
to move feature from Program Increment to backlog.
This allows to make a single request instead of 2 to move feature.
In future patches, this route will be improve to allow reorder in program
backlog (see story for details).

How to test:
 1)
   - Have features in program increment
   - Drag'n'drop feature from program increment to backlog
   |=> No expected changes. If you reload, the feature is at the right place

 2)
  - Have feature not in top backlog and not planned in PI
  - Go to the feature artifact view
  - Click on dropdown to move it to top backlog
  - Go to top backlog
  |=> There is the feature

 3)
  - Use masschange to move/remove features to top backlog
  - Go to top backlog
  |=> Features are moved/removed

Change-Id: I436bb72e965192f64659aa002e6d1b148237a437