Ability to switch scale to week
part of: story #19734 have a Roadmap widget on dashboard

Note to reviewers:
I use current-week-number to get the number of the week. There is no methods in the Date object to retrieve it, and the algorithm to compute it is quite complex
(years with 52/53 weeks, leap years and stuff).

How to test:
- switch the scale to "Weeks"
--> The timescale is switched to weeks

Change-Id: Iaaed3302563452d54e6012deda74eb1f9e3c61ec