Transform the artifact modal into a lib
Part of request #20062 Transform the artifact modal into a lib

How to test:
Run end-to-end tests for agiledashboard and testmanagement.

No functional change expected.

Notes:
The artifact modal is now an internal lib. The two PNG images included
are now inlined as base64. All webpack configuration, npm modules, mocks
and env variables previously needed to build the artifact modal should
have been removed from TTM and AD plugins.
It is bundled with webpack because I could not make it work with vite,
this can be done at a later contribution.

Change-Id: Ic9a6bce5b0ec131e56c0f16244f0585f9df0ffb8