Reorder elements in program backlog
This is part of story #20111 change features priority

How to test:
 - Have some features in program backlog
 - Reorder elements
 - Refresh

Expected results:
 - Elements keep their new places

Note to reviewer: There is no ability to order element when you are
moving it from PI to program backlog.
This will be done in next patch because there is a part of refactoring
that can be a little big and I didn't want to complicate this patch.

Change-Id: I711fadcf853759f15ac7d665ac24a339d4793c47