Refactoring: decrease number of nullable dates
A time period was depending on 3 different dates:
- the first "start" date of tasks
- the last "end" date of tasks
- today (to be able to display Today indicator)

Since tasks can have no start date or no end date, we had to pass
Date|null dates everywhere.

The current proposal moves the today date outside of time period
construction. The code and the tests are now simpler to read (I hope).

It is a technical refactoring, no functional changes are expected.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I2e945ace13a13ffbc50d58a8da5414bda8c1de19