Base of the PATCH PI REST route
Part of story #20111 change features priority

How to test:
- call PATCH /api/v1/program_increment/$id/content with a payload like
this:

  { "add": [{"id": <feature_id>}] }

- if the given PI $id does not exist, you will get a 404 error
- if the given PI $id is not a Program Increment (e.g. a User Story),
you will get a 404 error
- if the current user does not have permission to prioritize features,
you will get a 403 error
- if all conditions are satisfied, you will get a 501 error (Not
implemented). This is temporary and is meant to ease the review to not
produce huge patches.

There is no check on the "add" for now, they will be added later.

Note: I also moved some classes:
- move Planned Program Increment to Program Increment namespace, not
Feature Content
- move PI Not Found exception to Domain namespace
- move Retrieve PI interface to PI namespace, not content

Change-Id: Id3612c4ef72c1f6c0d195b1d47208eb12886de3c