Task bar should include end date.
We should ensure that the end date is part of the bar, else it produces
[strange results][0].

The issue is now visible with week timescale. It was barely visible in
month and quarter scale because the difference in pixel was too small.

If you have a task that starts on Sunday and ends on the following
Monday, the corresponding bar should have a width of two days:

```
         +----------+----------+
         |    W10   |    W11   |
+--------+----------+----------+
|        |          |          |
| #123   |        ┏━|━┓        |
| A task |        ┗━|━┛        |
|        |          |          |
+--------+----------+----------+
```

Similarly, if you have a task that starts on Monday and ends on the
following Monday, the corresponding bar should have a width of eight
days:

```
         +----------+----------+----------+
         |    W10   |    W11   |    W12   |
+--------+----------+----------+----------+
|        |          |          |          |
| #123   |          |┏━━━━━━━━━|━┓        |
| A task |          |┗━━━━━━━━━|━┛        |
|        |          |          |          |
+--------+----------+----------+----------+
```
Part of story #19734: have a Roadmap widget on dashboard

[0]: https://gerrit.tuleap.net/c/tuleap/+/22396/comment/677ae7c6_e6ea5c09/

Change-Id: I640eb781c1d92ebffb873dd38b513f20b31d8d97