Prepare the field for progress display
The display of a milestone was done by rotating a div. This had two
disadvantages:
* The look 'n feel was a bit blurry, due to the transform.
* The display of the progress bar in this case is complicated, you have
to counter the translation to see an horizontal progress.

This contribution targets both points, by using a svg element in place
of a rotated div.

By default the progress of tasks is set to 1 (100%), but you can test
with other figures (0, 0.3, …) in `RoadmapTasksRetriever.php`. The hard
coded value 1 will use the Progress semantic once it will be implemented
in a dedicated story.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I5c593b0ffda3635e57799df9feed1034f643ada6