Display years on top of time units
It might be difficult to know which year a month or a week is belonging
to, unless users hover to display the tooltip. Always displaying the
year seems to be more UX friendly.

Note: there isn't any separation between years for now (no border, no
alternate background, …). It will be done later, if needed, by Chief of
Design.

The today indicator pin has now its head above the time period,
according to the latest iteration of the mockup.

Part of story #19734: have a Roadmap widget on dashboard

Change-Id: I6190d422476d2897b0f392b9065cd66d0d7c66c4