Nullable progress of a task
When there is no semantic Progress defined on a tracker, then the
corresponding artifacts does not have any progress defined.

There will be a visual distinction between a task at 100% and a task
without progress:

* Task at 100%: a bar with a full progress bar with the text "100%"
* Task without progress: a bar with a full progress bar

(this will be done in a dedicated changeset, this commit prepares the
field)

No functional change.

Part of story #20919: define progress of artifacts

Change-Id: I113d72afc52c1a3327f329794ca1260337ce6641