Display percentage of progression
If task does not have any progress (null), then no percentage is
displayed.

If task have a percentage then it is displayed:
* next to the progress bar
* or inside the progress bar if there is not enough room
* or outside of the task bar if it is too small

The popover displays the percentage as well.

If task is a milestone, then percentage is always displayed outside of
the "bar".

If percentage is displayed outside of the bar, then the dependency
arrows start after the percentage instead of the bar so that text is
readable.

Note: to test this you will have to edit manually the file
`plugins/roadmap/include/Roadmap/REST/v1/RoadmapTasksRetriever.php` to
inject a progress between 0 and 1 (instead of null).

Part of story #20919: define progress of artifacts

Change-Id: Icbb8e4b51ba252f063f9b72c8348121dd726be0d