Refactoring: use HTTP header to check webhook type
This is part of request #20918 GitLab tags that contains a reference to Tuleap should be backlinked

Webhook content changed with breaking changes. "event_type" key was
renamed to "object_kind" and "event_name" key is not in all webhook
payload.

So we now use the request header "X-Gitlab-Event" to distinguish webhook
type.

Change-Id: I5b30eae01ce716a7d4edcf2d5ba67de07e1fbfc8