fix: Percentage display should not change color of the arrow
Given Task A from last week to yesterday,
And Task B from today to next week,
And a link from Task A to Task B
When the percentage of progression is displayed at the right of the
bar(*),
Then the color of the arrow kept is color (darkish, not redish) for
every time scales.

(*) You will need to hack the file `plugins/roadmap/include/Roadmap/REST/v1/RoadmapTasksRetriever.php`
to inject a progress between 0 and 1 (instead of null).

Part of story #20919: define progress of artifacts

Change-Id: I881d8ef2c7b3336ffce30132a35f61a153260a16