Extract FeatureRemovalProcessor
Part of story #20111 change features priority

No functional change expected.

This is part of a refactoring to extract parts of
ProcessTopBacklogChange into dedicated classes that can be reused in the
new PATCH program_increment/$id/content REST route.

Notes:
I introduce "Stub" classes for some of the interfaces. I'm going
to keep repeating the anonymous classes so I figured it was best to have
a dedicated Stub class.

The goal behind FeatureRemoval's named constructor is to force
developers to go through the checks in order to be able to call
FeatureRemovalProcessor. The only way to call it is to have a
FeatureRemoval, and the only way to build one is to make the
verifications.

Change-Id: I0e8c95e2321ea068506f6d5b86ae6cc809002f3f