Be able to navigate in Taskboard with arrow keys
Part of story #17555 : Have keyboard shortcuts in Taskboard

Introducing KeyboardShortcuts lib in Taskboard.

Swimlanes and cards can now take focus.
* If one is focused, up and down arrow will focus the previous and next one.
* If focus is on the last swimlane/card, down arrow will focus the first
swimlane/card.
* If focus is on the first swimlane/card, up arrow will focus the last
swimlane/card.

Right and left arrows will focus the first card of respectively the
right and left cell.
If focus is in the last or first cell, the first card of the last and
first cell will be respectively focused.
If cell is empty, the add button will take focus.

This makes it easier to reach the desired swimlane and card without having to tab
through all the elements of the previous ones.
Tabbing order remains the same, that way, after reaching a swimlane or card, user can tab and navigate inside it.

Change-Id: Id1228e174f9005ef41bde2dc3ada73236d1c32e6