Remove unused CampaignStatusChangeUnknownValueException class
This exception can never be thrown because the value we verify can only
be instantiated with those two values.
As a result the exception can never be thrown in any real life scenario.
This sort of situation is better handled at the type system level than
during runtime: no execution cost and we are certain the verification is
always done even if we reuse this value somewhere else.

This is part of request #20088 Improve the management of closed campaigns

Change-Id: I591b80aea3fa1d14f8a73b2372144705d15e31fa