GitLab tags are now referenced when a reference is added
This is part of request #20918 GitLab tags that contains a reference to Tuleap should be backlinked

The repository must have a token defined, otherwise the tag message
cannot be retrieved.
In addition, the webhook configured at GitLab side
must be updated to add "Tag push" if the configuration previously
exists. New integration will add by default the "Tag push".

If a tag is pushed and have one or multiple TULEAP-XXX in its message,
the references are now added.
The reference is displayed in the artifact view and is clickable.
No style has been done in this patch.

Change-Id: I228cb474de5970cc1c2a12e448b61fc00682ac38