Provide a way to easily launch outdated browsers that are still ~ supported
Developpers can now use 'old-firefox' and 'old-chromium' to launch
outdated versions of these browsers. Each time they are started, a new
profile is created to be sure they do not generate conflicts with the
day-to-day browsers of the developpers.

Those browsers are not put in the main dev tools derivation to avoid
loading them when the developpers does not need them. It is supposed
to be something that is needed occasionally, no need to waste more
bandwith and disk space than needed.

Closes request #20935: Bring best effort support back to Firefox 68 and Chrome 70

Change-Id: I3f5046aa90595eb475a1d456de0ae851aa0342a2