Refacto: Import builder in object class ProgramIncrementTracker
This is part of story #20111 change features priority

No expected changes.
You can test with adding tracker in Plan.

This patch aims to use builder in the class instead of another object.
This makes it possible to be sure that ProgramIncrementTracker is an existing tracker,
because some verification are done before the instantiaton.

Without this patch, we need to do some manually verification after/before/never instantiation.

In future developments, it will not be forgotten to check that
`tracker_id` does exist, before to create a ProgramIncrementTracker.

Change-Id: I57eac4f1d1dddae7ec11349c9ade3a4eb1b0ccae