fix: popover is not nicely displayed on small resolutions
When there is only one or two tasks displayed in the roadmap widgets,
there is not enough room to let popper.js (the underlying library that
is used for placements) display the popover.

Furthermore it has been brought to our attention that there
might be some flickering.

By upgrading the library [to its latest version][0], the display seems
te be a little bit better.

There is still weird behavior though, for example when a task bar is too
long for the scrolling area, the popover might be hidden. It will be
managed in a dedicated commit.

Part of story #19734: have a Roadmap widget on dashboard

[0]: https://popper.js.org/docs/v2/migration-guide/

Change-Id: Iac28e2b6f57e8a63295f2c4950eaa673aaf3c389