fix: typo in french translation for progression usage
In french we have this sentence displayed while we are managing the
progression semantic:

> À plus long terme, nous prévoyons de l'utiliser dans le Dashboard agile.

The capitalized "Dashboard" at the middle of the sentence seems weird.
It is not a french term, but at least it should be in lower case.

French translations found for Agile Dashboard in our .po file:
* Agile Dashboard
* tableau de bard agile
* dashboard agile

We should stick to one translation to not loose the user, but it
deserves its dedicated commit to handle that.

Part of story #20919: define progress of artifacts

Change-Id: Ib61c861ea2da83280cf4a5591840050974ddc42e