Part of request #20905 : Fix focused test in filtered tests list
User looses focus when using `filter remaining tests` and `mark test as`
shortcuts.

Given that, from the tests list:
* I use a shortcut to mark a test.
* I use the shortcut to filter it.
In those two cases, if the focused test is removed from the list due to
filtering, the problem occurs.

This patch aims at solving the first case.
`Mark test` shortcuts now jump to the next test in list after marking the test.

Also, some refactoring is done to change `document` call to a `doc` parameter.

The second part of the problem should be fixed in another patch.

Change-Id: Ia2c35bbec4e84a9f7aaa1937e47593099c7277a5