request #20978 Ease usage of jenkins for tuleap developments
Jenkins is now available at https://tuleap-web.tuleap-aio-dev.docker/jenkins
so it's easier to share the certificate that must be trusted between all
actors when you want tools to communicate with each other.

It's unlikely that existing Jenkins instances used by developers will
work out of the box after this change but it's not like it was actually
usable before.

Change-Id: I7f3351b01e58818a315322a4cd0ab6fc22ac59b1