Remove Mockery from OAuth2 Server plugin tests
This is mostly done as a way to give a reference in the codebase
of PHPUnit mock usages.

The plugin has been removed from Psalm analysis because Psalm is now
capable to do a correct analysis which result in a lot of errors
unrelated to this contribution. They will fixed as part of request #20106.

Part of request #21326: Favor PHPUnit mock system over Mockery

Change-Id: I89525ae1f149c31eed4c9261e115e05561902a74