refacor: compute is_milestone in Task
Instead of duplicating the algorithm to detect if a task is a milestone
or not, we now compute the information earlier, as soon as the task is
retrieve from the API.

It will be easier later if we need to move this information backend
side.

No functional change in the display of milestones.

Part of story #20979: Display children in Roadmap

Change-Id: Ic2e84e1cb6cfa3cc841a46cfe1638eed0877ee76