Refacto: UserCanPrioritize instead of plain \PFUser
This is part of story #20111 change features priority

This patch aims to use UserCanPrioritize object, instead of PFUser in
ProgramManagement plugin when it's possible.
UserCanPrioritize can only be used when we treated a program with a user and some checks were done.
It's not possible to change all PFUser to this new object.

No expected changes.

How to test:
 - Have 2 users:
   -> one can prioritize
   -> another not (anonymous for example)
 - Go to program service
   => With first user, you can move features
   => With anonymous, you can't

Change-Id: I4651add4a4a99a906491c2a652689f8527296b5e