feat: display caret next parent tasks
When a task have subtasks, then a caret is displayed to indicate to the
user that it can be expanded/collapsed.

For now only the caret is displayed. There is no expand/collapse, it
will come later.

In order to ease the development and not produce a too big changeset,
the detection of subtasks is stubbed in the retrieval of the tasks. As
soon as a task has _is_child dependencies, then we consider that it has
subtasks. This detection will be moved to backend side in a dedicated
changeset, and the REST route will return directly the information.

Part of story #20979: Display children in Roadmap

Change-Id: I12c664f8db2771c028f2ce9af3aba3f80def488b