fix: Rename roadmap_tasks/:id/children
We prefer to speak about subtasks. => roadmap_tasks/:id/subtasks

Rationale: "children" is too linked to the artifact link _is_child. We
don't know yet if a subtask will be another thing than an artifact
child. For example it might be based on the artifact link is_subtask,
or depends_on.

By choosing "subtasks" instead of "children" we introduce a specific
term that is more generic than the previous one. It will prevent a
breaking change in the API if the above asumptions become true.

Part of story #20979: Display children in Roadmap

Change-Id: Ice6c9a775d707fee3181bc97a000297c278170ac