fix: up/down left arrows should not be marked as danger
Have the following data:
Task A: 2021426 -> 20210503
Task B: 20210504 -> 20210506
Task A is linked to Task B

Display the widget. Display links.
=> For every time scales, the arrow should not be marked as danger.

The issue comes from the fact that when the task bar is too small, we
force it to have a minimum width to be visible (there is no point to
display a 1px bar).
However since the previous algorithm was using the width of the bar to
detect if an arrow should be marked as danger, in the aforementioned
scenario it was falsely displaying a danger arrow.

Now we use dates of the tasks so it should be better.

Closes request #21328: Dependency arrows can be falsely displayed as red

Change-Id: If44057a0bdb2cb7c36ee799aedeccac533abe40a