Force "strict" PHPUnit mocks
This contribution makes that all methods called on mocks created by
PHPUnit are explicitely defined (same constraint that we currently have
with Mockery).

This is done by creating a TestCase specific to Tuleap and enforcing its
usage via a PHPCS rule. This also enforced on REST and SOAP tests even
if they are not concerned by the usage of mocks just to prevent some
confusion ("when should I use this one instead of the other?").

Part of request #21326: Favor PHPUnit mock system over Mockery

Change-Id: Ib69dc2bf2098240f120f0d1b7a1996aa31a859bb