Gain performance when planning Feature in PI
This is part of story #20111 change features priority

For the moment, in the quest to reduce time when planing Feature in
Program Increment, there is a first patch that helps improve time perfomance.
I know it's not the best performance gain, but that permits to remove 100 SQL
queries.

My tests were done with this configuration:
 - 1 Program Increment in Program project
 - 3 Teams project linked to the Program
 - 3 Mirrored Release linked to the Program Increment
 - 7 User Stories in each Mirrored Release linked to the Feature as child

Some results with BlackFire:
 -> Without the patch:
    => Timing: 1.87s
    => CPU: 1.38s
    => SQL: 682 ms / 951 requests

 -> With this patcg
    => Timing: 1.78s
    => CPU: 1.33s
    => SQL: 649 ms / 839 requests

How to reproduce:
 - Have the same confirugation
 - Go to Program Service
 - Plans feature in PI

Change-Id: I61eefdafbdab43a55677ae79e20822d2f9f7dda2