Introduce the 'Resolve' close artifact keyword
Part of story #17183 close artifacts from commit messages

In this step, only a new comment is posted in the referenced Artifacts
from the GitLab commit.
The artifact is not closed or workflow transition are not taken in
account.
Only the 'resolve' or 'resolves' keys are supported.

How to test:
 - Tracker related to artifact 1000 and 108 must not have Status
   semantic defined
 - In the commit message add some Tuleap references with the 'resolve' key before.
e.g : This is a commit with references: Tuleap-95 , resolve tuleap-1000,
resolves Tuleap-108

=> In that example, the artifacts 108 and 1000 should have a new follow
up comment with the Tuleap user (if found) reference or with the commit
author.

Change-Id: I579a640d05c7d58e4670097b2a01d1d10f4d8299