Reference git_tag is taken into account
This is part of request #21327 Cross references in Tuleap Git tags should be extracted

When you add reference like git_tag #repo_name/tag_name in an artifact
comment for instance, the reference is now usable.

If an already existing and used project reference git_tag is in the
project, this reference will be used instead of the system reference.

Change-Id: I959a000038db5852b9b72fd89be0c22fb7da33e4