feat: loading state when displaying subtasks
When a task has subtasks and users click on the caret to expand it, then
skeleton rows are displayed under the task.

If there was links displayed, then they are not impacted by the new rows
that are displayed.

If users click again on the caret, then the skeleton rows disappear.

Part of story #20979: Display children in Roadmap

Change-Id: Ic1f9f38d4a00afb23c5d6ef3041b96dacd0da0b7