request #21333: @tuleap/vue-breadcrumb-privacy uses a built-in HTML element as component ID when built in prod mode
The issue is visible when the library is built in prod mode but the Tuleap core scripts
are built in dev mode (otherwise the warning is silenced).

Trace:
```
[Vue warn]: Do not use built-in or reserved HTML elements as component id: u
```

The issue is caused indirectly by the minifier that will attempt to
squeeze a few bits by changing the class name. Since we are using
vue-class-component the default ID attributed to the component is the
class name. To avoid any issue we now explicitely set the name of the
component.

Change-Id: Ife8494b5f02604ec1e4a2d904d4d813922a3113e