Make REST tests easier to troubleshoot
When something goes wrong with REST tests it's hard to
troublshoot, mainly because there are no logs easily
accessible and because of custom feed of the DB. This
commit fix this.

As an interesting side effect, there is a need to update
cardwall configuration in 'private-member' project
because there SetupMysql command sets the Tuleap version
in base and it's used by application code to detect if
we activate cardwall or not. Believe it or not, it was
a huge moment of fun to troubleshoot that (not).

Part of request #21334 enalean/tuleap-aio should be easier to test

Change-Id: I64b8545fdbfdb9da59a6ed37d60c044448e63161