feat: display subtasks
When the user expands a task, its subtasks are loaded and displayed.
(skeletons are still displayed while loading).

If there are links between subtasks or between task and subtasks, then
they are displayed.

The timeperiod is adapted according to the displayed set of tasks and
subtasks.

Error management will be done later in a dedicated contribution.

Part of story #20979: Display children in Roadmap

Change-Id: Ie75f76ac8f3f4c2b8d447512ac1b92037a770d3c