feat: error state for subtasks
If the retrieval of subtasks fails, then an error message is displayed.
For now the message is generic, we will see in the wild if we need to
give more details to the end user.

Implementation note 1:
Using directly the tlp-alert component brings too much overriding,
therefore we now expose the component colors so that we have better
control of the message display inside the gantt chart.

Implementation note 2:
Due to the way the gantt board is built (scrolling areas, header fixed,
…), the dom element containing the message is written outside its logical
container. The latter takes the necessary space (task's height) while
the former is positionned absolutely at the right position.

Implementation note 2:
The v-bind:key used for tasks has been changed to a more accurate key.
The key should be linked to the element, not its position in the
collection, else VueJS will be confused while diffing/rendering
components when the collection is changed. This is highlighted here with
popover that was targeting the wrong task bar.

Part of story #20979: Display children in Roadmap

Change-Id: I10f29e7e56409b8fbaf0bf20e0bb5ab44f05aa45