feat: Display project for children
Given StoryA is children of EpicB,
And they don't belong to the same project,
When we expand the subtasks of EpicB,
Then the project label of StoryA is displayed next to its xref.

Note: a Design intervention needs to be done later to adjust the display
when the project label is too long. Currently the xref is disappearing
when it is the case ☹️

Part of story #20979: Display children in Roadmap

Change-Id: Ied486bf66ba5e3b1980cd507ac6dd4c410bb5b65