request #21345: Remove embedded jstimezonedetect library
The library is no more maintained upstream (in fact upstream does not
exist anymore) and the same thing can be achieved with native APIs now.

Account_TimezonesCollection has been updated to use the timezone known
by PHP to make sure we are always up to date with the upstream IANA
timezone database so we have a better chance to select the correct
timezone client side.

Change-Id: Idd23c5daf91aebd51fc18939a85fd23570ec1f58