feat: Display subtasks empty state
Given EpicA has every children that cannot be displayed as tasks, when the
user expands EpicA, then an info message is displayed to inform the user
about the impossibility to display them.

A child cannot be displayed as task if at least one of those conditions
is true:
* The user don't have permission to see it
* Its tracker does not have a title semantic
* Its tracker does not have a timeframe semantic
* It does not have start date nor end date
* It has a start date > end date

For performance reasons not all conditions above can be detected when we
retrieve parent tasks information (or situation might evolve between the
display of the widget and the expand of the task), therefore the
presence of the caret might be misleading. If it appears that a task
don't have any subtasks to display, then a button [ok, got it] is
displayed so that the user can confirm undestood that there won't be
anything displayed, and the caret is then removed.

In a future contribution, we will try to detect earlier aforementionned
conditions in order to avoid as much as possible the display of a
useless caret.

Part of story #20979: Display children in Roadmap

Change-Id: I4243a21ffd0dca1c49e103b2bc0179f74e9d8122