feat: Remove dependencies w/o title|timeframe semantic
If an artifact linked to a task has no title or timeframe semantic, then
it will never be displayed in the gantt chart (as a task or as a
subtask). Therefore there is no need to return it in the collection of
dependencies.

While it had no impact on the display of links, there was visible
consequences for the display of children: a caret was displayed even if
no children could be displayed.

Even if the current proposal does not fix every edges cases, it has the
merit to prevent the display of the caret for the more common situation:
a child tracker without needed semantic. Other scenarii like the lack of
needed permission to read the artifact, or an artifact with start date >
end date is more complex and does not worth it for the Roadmap MVP.

Part of story #20979: Display children in Roadmap

Change-Id: I9d3da98dadb7e5f24947932fce6720b359f32ff0